src/ns1-ddns/app/requirements.txt
click
git+https://github.com/ns1/ns1-python.git
pyyaml
requests

